Unrolling the loop in Base.Sort.radix_sort_pass! increases performance by about 50%.
(on my Ryzen 3900x)

This makes the code somewhat less clean, but in my opinion the performance benefit is worth it.

Disclaimer: I did not fully build Julia including this PR (yet). The change is very small and I did test the modified function, but I do not want to setup everything to build Julia if the PR might be rejected based on a choice of code clarity vs. performance.

[Here is a simple test script](https://github.com/LSchwerdt/MiscJulia/blob/4c2c9c3abe04bcb5dd9cdcadb841ed64442958fd/radix_pass_unroll/radixpassunroll.jl) to try out the changed function in isolation.
This is the output:
```
2000/20000
...
20000/20000

base:     5.063 seconds
unrolled: 3.052 seconds
speedup factor: 1.66
Test Passed
```