@LilithHafner 2023-01-30
Thanks!

> This makes the code somewhat less clean, but in my opinion the performance benefit is worth it.

50% speedup for 20 lines of code is definitely worth it!

I wish the Julia compiler could do this automatically and I'd like to add a `#TODO: remove this once the compiler is smart enough to do it automatically`, but in the interim (which might be a very long time) I support this PR for its immediate performance gains.

I'm having trouble reproducing your performance results. On my computer (2019 MacBook Air), pasting your script into a Julia REPL results in the following output:

```
base:     7.235 seconds
unrolled: 6.829 seconds
speedup factor: 1.06
```

<details>

<summary> I also tried directly measuring calls to `sort!`, but found no change with this script:</summary>

```julia
using BenchmarkTools, Random

function benchmark(k = 8)
    for n in 2:k-2
        println(n)
        v = rand(10^n)
        @btime sort!($v) setup=(rand!($v)) evals=1 gctrial=false gcsample=false samples=10^(k-n);
        v = rand(Int, 10^n)
        @btime sort!($v) setup=(rand!($v)) evals=1 gctrial=false gcsample=false samples=10^(k-n);
    end
end

benchmark()

@eval Base.Sort function radix_sort_pass!(t, lo, hi, offset, counts, v, shift, chunk_size)
    mask = UInt(1) << chunk_size - 1  # mask is defined in pass so that the compiler
    @inbounds begin                   #  ↳ knows it's shape
        # counts[2:mask+2] will store the number of elements that fall into each bucket.
        # if chunk_size = 8, counts[2] is bucket 0x00 and counts[257] is bucket 0xff.
        counts .= 0
        for k in lo:hi
            x = v[k]                  # lookup the element
            i = (x >> shift)&mask + 2 # compute its bucket's index for this pass
            counts[i] += 1            # increment that bucket's count
        end
        counts[1] = lo                # set target index for the first bucket
        cumsum!(counts, counts)       # set target indices for subsequent buckets
        # counts[1:mask+1] now stores indices where the first member of each bucket
        # belongs, not the number of elements in each bucket. We will put the first element
        # of bucket 0x00 in t[counts[1]], the next element of bucket 0x00 in t[counts[1]+1],
        # and the last element of bucket 0x00 in t[counts[2]-1].

		#loop unrolled 4x
        k = lo
        while k <= hi - 4
            Base.Cartesian.@nexprs 4 _ -> begin
                x = v[k]                  # lookup the element
                i = (x >> shift)&mask + 1 # compute its bucket's index for this pass
                j = counts[i]             # lookup the target index
                t[j + offset] = x         # put the element where it belongs
                counts[i] = j + 1         # increment the target index for the next
                k += 1                    #  ↳ element in this bucket
            end
        end
        while k <= hi
            x = v[k]                  # lookup the element
            i = (x >> shift)&mask + 1 # compute its bucket's index for this pass
            j = counts[i]             # lookup the target index
            t[j + offset] = x         # put the element where it belongs
            counts[i] = j + 1         # increment the target index for the next
            k += 1                    #  ↳ element in this bucket
        end
    end
end

benchmark()
```

Output:

```
2
  1.779 μs (1 allocation: 896 bytes)
  1.532 μs (1 allocation: 896 bytes)
3
  15.251 μs (2 allocations: 10.12 KiB)
  15.976 μs (2 allocations: 10.12 KiB)
4
  180.667 μs (3 allocations: 86.36 KiB)
  199.008 μs (3 allocations: 86.36 KiB)
5
  1.866 ms (3 allocations: 789.48 KiB)
  2.106 ms (3 allocations: 789.48 KiB)
6
  23.867 ms (3 allocations: 7.64 MiB)
  27.281 ms (3 allocations: 7.64 MiB)
2
  1.794 μs (1 allocation: 896 bytes)
  1.567 μs (1 allocation: 896 bytes)
3
  15.819 μs (2 allocations: 10.12 KiB)
  16.955 μs (2 allocations: 10.12 KiB)
4
  180.848 μs (3 allocations: 86.36 KiB)
  192.534 μs (3 allocations: 86.36 KiB)
5
  1.778 ms (3 allocations: 789.48 KiB)
  2.069 ms (3 allocations: 789.48 KiB)
6
  23.991 ms (3 allocations: 7.64 MiB)
  27.122 ms (3 allocations: 7.64 MiB)
```

</details>

Time it takes to `sort!` a vector

Input | Before | After
---|---|---
rand(10^2) | 1.779 μs | 1.794 μs
rand(Int, 10^2) | 1.532 μs | 1.567 μs
rand(10^3) | 15.251 μs | 15.819 μs
rand(Int, 10^3) | 15.976 μs | 16.955 μs
rand(10^4) | 180.667 μs | 180.848 μs
rand(Int, 10^4) | 199.008 μs | 192.534 μs
rand(10^5) | 1.866 ms | 1.778 ms
rand(Int, 10^5) | 2.106 ms | 2.069 ms
rand(10^6) | 23.867 ms | 23.991 ms
rand(Int, 10^6) | 27.281 ms | 27.122 ms