@LilithHafner 2023-01-30
I doubt this will reveal much of note before https://github.com/JuliaCI/BaseBenchmarks.jl/pull/305 merges, but giving it a go anyway:

@nanosoldier `runbenchmarks("sort", vs=":master")`