@LSchwerdt 2023-01-30
`#TODO: remove this once the compiler is smart enough to do it automatically`
Yeah, I think you are right. It is not so much the amount of code, it is the feeling that it is somehow not necessary, because the compiler should be able to to this.

I did some more benchmarks using my test script on different hardware. It seems the speedup is not really reproducible. Maybe there is something wrong with my system? I think this should not be merged, unless someone else can reproduce the performance gain.

`n = 10^3, nRepeats = 10^8`
|  Hardware | time (base) | time (unrolled) | speedup factor |
| ------------- | ------------- | ------------- | ------------- |
| AMD 3900X | 156.316s | 164.073s | 0.95 |
| Intel 6700HQ | 210.218s | 243.852s | 0.86 |
| Intel 7820X | 174.933s | 193.171s | 0.91 |

` n = 10^7, nRepeats = 10^3`
|  Hardware | time (base) | time (unrolled) | speedup factor |
| ------------- | ------------- | ------------- | ------------- |
| AMD 3900X | 31.560s | 20.221s | 1.56 |
| Intel 6700HQ | 67.730s | 67.554s  | 1.00 |
| Intel  7820X | 38.566s | 39.933s | 0.97 |