@nanosoldier 2023-01-30
[Your benchmark job](https://github.com/JuliaLang/julia/pull/48459#issuecomment-1409072236) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/8653d59_vs_a1c4d85/report.md).