@vtjnash 2023-01-30
> 50% speedup for 20 lines of code is definitely worth it!

The extra icache pressure can make this undesirable, even if it performs well in benchmarks, so it is not always better for the compiler to do this automatically where the user did not ask for it.